fix: cast increment_vote return to int for bigint columns
The articles.likes/dislikes columns are bigint (Supabase's default
integer type), but the function declared its return as int, so the RPC
failed at runtime with 42804 (returned bigint != expected integer).
Cast the returned columns to int to match the declared signature.

+++ supabase/increment_vote.sql
@@ -35,6 +35,8 @@ begin
   end if;
 
   return query
-    select a.likes, a.dislikes from articles a where a.article_id = p_article_id;
+    select a.likes::int, a.dislikes::int
+    from articles a
+    where a.article_id = p_article_id;
 end;
 $$;
